requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
dask==2021.10.0
fastparquet==0.4.0
#Fiona==1.8.13.post1
fsspec==0.7.4
future==0.18.2
geopandas==0.7.0
idna==2.9
Landsat578==0.5.1
llvmlite==0.32.1
locket==0.2.0
lxml==4.9.1
munch==2.5.0
numba==0.49.1
#numpy==1.22.0
pandas==1.0.4
partd==1.1.0
pyarrow==0.17.1
pyproj==2.6.1.post1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4
requests==2.23.0
Shapely==1.7.0
six==1.15.0
thrift==0.13.0
toolz==0.10.0
urllib3==1.26.5